requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pandas==2.1.0
sqlalchemy==2.0.20
tqdm==4.66.0